+++ src/js/visual/configurator-csr.js
@@ -91,6 +91,7 @@ ripe.ConfiguratorCsr.prototype.init = function() {
     this.animations = [];
     this.isChangeFrameAnimationRunning = false;
     this._pendingOps = [];
+    this._postRenderCallback = null;
 
     // CSR variables
     this.rendererOptions = null;
@@ -526,6 +527,22 @@ ripe.ConfiguratorCsr.prototype.prcFrame = async function() {
     const positionRounded = Math.round(position);
 
     return `side-${positionRounded}`;
+};
+
+/**
+ * Sets the callback function for the post render call.
+ *
+ * @param {Function} callback The function to be called.
+ */
+ripe.ConfiguratorCsr.prototype.setPostRender = function(callback) {
+    this._postRenderCallback = callback;
+};
+
+/**
+ * Clears the post render callback registry.
+ */
+ripe.ConfiguratorCsr.prototype.unsetPostRender = function() {
+    this._postRenderCallback = null;
 };
 
 /**
@@ -1200,6 +1217,7 @@ ripe.ConfiguratorCsr.prototype._render = function() {
     if (!this.camera) throw new Error("Camera not initiated");
     if (!this.renderer) throw new Error("Renderer not initiated");
     this.renderer.render(this.scene, this.camera);
+    this._onPostRender();
 };
 
 /**
@@ -1271,6 +1289,13 @@ ripe.ConfiguratorCsr.prototype._onAnimationLoop = function(self) {
 
     // renders a frame
     self._render();
+};
+
+/**
+ * @ignore
+ */
+ripe.ConfiguratorCsr.prototype._onPostRender = function() {
+    if (this._postRenderCallback) this._postRenderCallback();
 };
 
 /**
@@ -1393,6 +1418,9 @@ ripe.ConfiguratorCsr.prototype._onPostConfig = async function(self, config) {
         // checks if initials are enabled
         const initialsEnabled = this.owner.hasPersonalization();
 
+        // gets the initials config
+        const initials = await this.owner.getInitialsConfigP();
+
         // checks if it should load assets used by the initials
         let baseTexturePath = null;
         let displacementTexturePath = null;
@@ -1467,8 +1495,7 @@ ripe.ConfiguratorCsr.prototype._onPostConfig = async function(self, config) {
             }
         }
 
-        // gets the initials and initials.3d set from the config
-        const initials = config.initials || {};
+        // gets the initials.3d set from the config
         const initials3d = initials["3d"] || {};
 
         // unpacks initials options